refactor: split into files + lint
