+++ packages/cozy-device-helper/.eslintrc
@@ -1,0 +1,4 @@
+{
+  "env": { "jasmine": true },
+  "extends": ["cozy-app"]
+}
