+++ packages/cozy-device-helper/cordova.js
@@ -1,0 +1,2 @@
+// cordova
+export const isCordova = () => window.cordova !== undefined
