+++ packages/cozy-device-helper/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "prepublishOnly": "yarn run build",
     "build": "babel index.js -o build.js --minified",
-    "test": "yarn run jest"
+    "test": "yarn run jest",
+    "lint": "yarn eslint --fix '*.js'"
   },
   "repository": {
     "type": "git",
@@ -19,12 +20,12 @@
   },
   "homepage": "https://github.com/cozy/cozy-libs/blob/master/README.md",
   "devDependencies": {
+    "babel-cli": "6.26.0",
     "babel-core": "6.26.3",
     "babel-jest": "23.0.1",
-    "babel-preset-cozy-app": "^0.10.1",
+    "babel-preset-cozy-app": "0.8.0",
     "babel-preset-env": "1.7.0",
-    "babel-cli": "6.26.0",
-    "babel-preset-cozy-app": "0.8.0"
+    "eslint-config-cozy-app": "^0.10.0"
   },
   "dependencies": {
     "lodash": "4.17.10"
